requirements.txt
fastapi
uvicorn 
gunicorn 
pydantic-settings 
asyncio 
torch 
transformers 
accelerate 
bitsandbytes
googletrans 
python-multipart 
pillow